requirements.txt
# PlantCare AI – Python dependencies
# Core Flask backend
flask>=3.0

# Deep learning / model inference
tensorflow>=2.15
numpy>=1.24

# Image processing
opencv-python>=4.9
Pillow>=10.0

# Text-to-Speech (voice output)
gTTS>=2.5

# Dataset download (Kaggle)
kagglehub>=0.3
<<<<<<< HEAD

# Google Gemini AI (disease analysis + translation) — new SDK
google-genai>=1.0.0

# Translation (Telugu / Hindi)
deep-translator>=1.11.4

# .env file support for API keys
python-dotenv>=1.0.0
=======
google-generativeai>=0.5.0
deep-translator>=1.11.4
>>>>>>> 5c48b9065574dd81b6b9f74d3debcd7d6eed8a31
